# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Market API
  description: 生鲜电商系统 RESTful API 规范
  version: 1.0.0-SNAPSHOT
  contact:
    name: Fresh Market Team
    email: contact@freshmarket.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发环境
  - url: https://api.freshmarket.com/api/v1  
    description: 生产环境
paths:
...
  /orders:
    post:
      summary: 创建订单
      description: 根据购物车内容或指定商品创建订单
      operationId: createOrder
      tags:
        - 订单管理
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 订单创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 创建订单失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: 获取用户订单列表
      description: 分页获取当前用户的订单列表
      operationId: getUserOrders
      tags:
        - 订单管理
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码(从0开始)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - name: status
          in: query
          description: 订单状态筛选
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedOrderResponse'
  /orders/{id}:
    get:
      summary: 获取订单详情
      description: 根据订单ID获取订单详细信息
      operationId: getOrder
      tags:
        - 订单管理
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 订单ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '404':
          description: 订单不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/pay:
    post:
      summary: 支付订单
      description: 对指定订单进行支付处理(模拟)
      operationId: payOrder
      tags:
        - 订单管理
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: 订单ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayOrderRequest'
      responses:
        '200':
          description: 支付成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: 支付失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# =============================================================================
# 数据模型定义
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误码
          example: "INVALID_REQUEST"
        message:
          type: string
          description: 错误信息
          example: "请求参数不正确"
        timestamp:
          type: string
          format: date-time
          description: 错误发生时间
        path:
          type: string
          description: 请求路径
          example: "/api/v1/users/register"
      required:
        - code
        - message
        - timestamp
    PageInfo:
      type: object
      properties:
        page:
          type: integer
          description: 当前页码
          example: 0
        size:
          type: integer
          description: 每页数量
          example: 20
        totalElements:
          type: integer
          format: int64
          description: 总记录数
          example: 100
        totalPages:
          type: integer
          description: 总页数
          example: 5
        hasNext:
          type: boolean
          description: 是否有下一页
        hasPrevious:
          type: boolean
          description: 是否有上一页

    # 用户相关模型
...
    CreateOrderRequest:
      type: object
      properties:
        addressId:
          type: integer
          format: int64
          description: 收货地址ID
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          description: 订单商品(为空时使用购物车商品)
        remark:
          type: string
          description: 订单备注
          maxLength: 200
          example: "请送货上门"
      required:
        - addressId
    OrderItemRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: 商品ID
          example: 1
        quantity:
          type: integer
          description: 购买数量
          minimum: 1
          maximum: 999
          example: 2
      required:
        - productId
        - quantity
    OrderStatus:
      type: string
      enum: [CREATED, WAITING_PAYMENT, PAID, CANCELLED]
      description: 订单状态
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 订单项ID
          example: 1
        productId:
          type: integer
          format: int64
          description: 商品ID
          example: 1
        productName:
          type: string
          description: 商品名称
          example: "有机苹果"
        productImage:
          type: string
          format: uri
          description: 商品图片
          example: "https://example.com/images/apple.jpg"
        price:
          type: number
          format: decimal
          description: 商品单价
          example: 12.50
        quantity:
          type: integer
          description: 购买数量
          example: 2
        subtotal:
          type: number
          format: decimal
          description: 小计金额
          example: 25.00
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 订单ID
          example: 1001
        orderNo:
          type: string
          description: 订单号
          example: "ORD202412051001"
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
          format: decimal
          description: 订单总金额
          example: 125.50
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        remark:
          type: string
          description: 订单备注
          example: "请送货上门"
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OrderResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItemResponse'
            address:
              type: object
              properties:
                name:
                  type: string
                  example: "张三"
                phone:
                  type: string
                  example: "13812345678"
                address:
                  type: string
                  example: "北京市朝阳区xxx路xxx号"
    PagedOrderResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'
        page:
          $ref: '#/components/schemas/PageInfo'
    PayOrderRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [WECHAT_PAY, ALIPAY, BANK_CARD]
          description: 支付方式
          example: "WECHAT_PAY"
      required:
        - paymentMethod
    PaymentResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          description: 订单ID
          example: 1001
        paymentId:
          type: string
          description: 支付流水号
          example: "PAY202412051001001"
        status:
          type: string
          enum: [SUCCESS, FAILED, PENDING]
          description: 支付状态
          example: "SUCCESS"
        paymentMethod:
          type: string
          description: 支付方式
          example: "WECHAT_PAY"
        amount:
          type: number
          format: decimal
          description: 支付金额
          example: 125.50
        paidAt:
          type: string
          format: date-time
          description: 支付时间